+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Add the FilesModel::findMultipleFoldersByFolder() method (see contao/core#7942).
  * Add the fragment registry and the fragment renderers (see #700).
  * Use a double submit cookie instead of storing the CSRF token in the PHP session (see #1065).
  * Decorate the Symfony translator so it reads the Contao labels (see #1072).